refactor: extract shared isPublished helper, inline readVersion, remove ext alias
